feat: added a new field to ProductItem
Added itemDescription(String) field

feat: added ItemAddToCartMenu, ItemCard, ItemIcon, ItemsAddToCartMenuCarousel, ProductItemCounter

ItemAddToCartMenu is a composable which provides the functionality for the ProductItem to be added to the cart. This composable is built with additional composables such as ItemCard, ItemIcon, ItemsAddToCartMenuCarousel, ProductItemCounter.

ProductItemCounter composable provides us with the items-to-be-added-to-cart functionality.

ItemsAddToCartMenuCarousel is a row with the ItemIcon components representing other items.

ItemIcon is a composable representing ProductItem image and link to the corresponding item.

feat: added the auxiliary images and string